+++ src/XUMM/Callback/PaymentCallback.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace XummForWoocomerce\XUMM\Callback;
+
+class PaymentCallback
+{
+    public function __construct()
+    {
+
+    }
+}
