+++ src/XUMM/Callback/SignInCallback.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace XummForWoocomerce\XUMM\Callback;
+
+class SignInCallback
+{
+    public function __construct()
+    {
+
+    }
+}
