+++ src/XUMM/Callback/TrustSetCallback.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace XummForWoocomerce\XUMM\Callback;
+
+class TrustSetCallback
+{
+    public function __construct()
+    {
+
+    }
+}
